test: lastSideEffectCounter
